fix issue #1597
