+++ avalanche/benchmarks/scenarios/__init__.py
@@ -8,3 +8,4 @@ from .task_aware import *
 from .dataset_scenario import *
 from .exmodel_scenario import *
 from .online import *
+from .validation_scenario import *
