+++ tests/training/test_plugins.py
@@ -16,6 +16,8 @@ from torch.utils.data.dataloader import DataLoader
 from avalanche.benchmarks import (
     nc_benchmark,
     GenericCLScenario,
+)
+from avalanche.benchmarks.scenarios.validation_scenario import (
     benchmark_with_validation_stream,
 )
 from avalanche.benchmarks.utils.data_loader import TaskBalancedDataLoader
